backend/app/requirements.txt
fastapi[standard]>=0.121.3 
uvicorn[standard]>=0.38.0
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
ultralytics >=8.3.227
numpy >=2.2.6
pillow >=12.0.0 
pandas >=2.3.3
scikit-learn >=1.7.2
matplotlib >=3.10.6
python-multipart >=0.0.20